+++ CHANGELOG.rst
@@ -15,9 +15,17 @@ Change Log
 
 Unreleased
 ----------
-[4.6.11]
+[4.6.13]
 -------
 feat: added fields for holding encrypted data in database (ENT 5613)
+
+[4.6.12]
+--------
+feat: unlink degreed2 inactive user
+
+[4.6.11]
+--------
+chore: Aligning SAP naming conventions
 
 [4.6.10]
 --------